+++ src/main/java/com/example/publicdatabackend/service/UsersService.java
@@ -4,12 +4,9 @@ import com.example.publicdatabackend.domain.restaurant.Restaurant;
 import com.example.publicdatabackend.domain.reviews.Reviews;
 import com.example.publicdatabackend.domain.users.Users;
 import com.example.publicdatabackend.domain.users.WishListRestaurant;
-import com.example.publicdatabackend.dto.RestaurantDto;
-import com.example.publicdatabackend.exception.UsersException;
+import com.example.publicdatabackend.dto.restaurant.RestaurantDto;
 import com.example.publicdatabackend.repository.ReviewsRepository;
-import com.example.publicdatabackend.repository.UsersRepository;
 import com.example.publicdatabackend.repository.WishListRestaurantRepository;
-import com.example.publicdatabackend.utils.ErrorResult;
 import com.example.publicdatabackend.utils.ExceptionUtils;
 import com.example.publicdatabackend.utils.RestaurantDtoConverterUtils;
 import lombok.RequiredArgsConstructor;
