+++ src/test/java/com/example/publicdatabackend/service/test/RestaurantServiceTest.java
@@ -1,7 +1,7 @@
 package com.example.publicdatabackend.service.test;
 
 import com.example.publicdatabackend.domain.restaurant.Restaurant;
-import com.example.publicdatabackend.dto.RestaurantDto;
+import com.example.publicdatabackend.dto.restaurant.RestaurantDto;
 import com.example.publicdatabackend.repository.KakaoReviewsRepository;
 import com.example.publicdatabackend.repository.RestaurantRepository;
 import com.example.publicdatabackend.repository.ReviewsRepository;
